Explorar con widget separado como loading_icon.dart y el explorar_screen.dart esta cambiado
