+++ appanimals/lib/screens/explorar_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../widgets/loading_icon.dart';
 
 class ExplorarScreen extends StatefulWidget {
   const ExplorarScreen({Key? key}) : super(key: key);
@@ -99,29 +100,11 @@ class _ExplorarScreenState extends State<ExplorarScreen> {
               Positioned(
                 bottom: 40,
                 left: size.width * 0.5 - 30,
-                child: const _LoadingIcon(),
+                child: const LoadingIcon(),
               ),
           ],
         ),
       ),
-    );
-  }
-}
-
-class _LoadingIcon extends StatelessWidget {
-  const _LoadingIcon({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(10),
-      height: 60,
-      width: 60,
-      decoration: BoxDecoration(
-        color: Colors.white.withOpacity(0.9),
-        shape: BoxShape.circle,
-      ),
-      child: const CircularProgressIndicator(color: Colors.blueAccent),
     );
   }
 }
